Updating for September 2017 files
This is going to take a bit of work. Looks like there are some problems with duplicates in the newer files that we need to sort out. Also, there seems to be some manual manipulation of the January 2017 file.

I am also renaming some of the variables so they make more sense to me, so there will be fairly substantial changes by the time it is done.